Int attributes are actually using Int64 according to the proto format for the default value.

+++ generate_wrappers.py
@@ -119,7 +119,7 @@ _SWIFTIFIED_TYPES = {
 }
 
 _SWIFTIFIED_ATTR_TYPES = {
-    'int': 'Int32',
+    'int': 'Int64',
     'float': 'Double',
     'bool': 'Bool',
     'string': 'String',
